+++ packages/util/src/index.ts
@@ -3,6 +3,7 @@ export * from './code-block';
 export * from './constants';
 export * from './equal';
 export * from './error';
+export * from './error-with-cause';
 export * from './fs';
 export * from './logger';
 export * from './nullish';
